leave physicalResistance an int. Use high bits to store a chance of breaking. Do condom rip test in ignoreCondom() renamed to testCondomRipped() since this is called after it is established the passed gamecharacter is a condom wearer, there is no reason to test for penis/condom. Include cum production in chance whether condom breaks. change the BROKEN_CONDOM itemeffect to GENERIC_BROKEN so it can be used for other purposes.

+++ src/com/lilithsthrone/controller/eventListeners/InventoryTooltipEventListener.java
@@ -867,7 +867,7 @@ public class InventoryTooltipEventListener implements EventListener {
 		// Attribute modifiers:
 		tooltipSB.append("<div class='container-full-width'>"
 				+ "<div class='container-half-width titular' style='width:calc(66.6% - 16px);'>");
-		float res = absClothing.getClothingType().getPhysicalResistance();
+		int res = absClothing.getClothingType().getPhysicalResistance();
 		tooltipSB.append(
 				(res>0
 					?"[style.boldGood(+"+absClothing.getClothingType().getPhysicalResistance()+")]"
